# contracts/kiln-http-api-v1.openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Kiln HTTP API",
    "version": "1.0.0",
    "description": "Canonical inventory of every production HTTP route, method, transport, handler owner, request type, response type, status, media type, path/header/query input, and structured error envelope. Every named payload component delegates its complete field contract to a canonical generated schema."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8420"
    }
  ],
...
  "paths": {
...
    "/v1/terminal/status": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Terminal status",
        "operationId": "terminal_terminal_status_get_v1_terminal_status",
        "x-kiln-handler": "terminal::terminal_status",
        "x-kiln-rust-signature": "async fn terminal_status() -> Json<TerminalStatusResponse>",
        "x-kiln-transport": "http",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalStatusResponse"
                }
              }
            },
            "x-kiln-rust-type": "TerminalStatusResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/terminal/ws": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Upgrade to the trusted embedded terminal WebSocket",
        "operationId": "terminal_terminal_ws_get_v1_terminal_ws",
        "x-kiln-handler": "terminal::terminal_ws",
        "x-kiln-rust-signature": "async fn terminal_ws( ws: WebSocketUpgrade, axum::extract::State(state): axum::extract::State<AppState>, headers: HeaderMap, ) -> impl IntoResponse",
        "x-kiln-transport": "websocket",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": false,
            "description": "Browser origin checked by the trusted terminal gate.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Successful response.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/WebSocketUpgrade"
                }
              }
            },
            "x-kiln-rust-type": "WebSocketUpgrade"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "ApiError": {
        "description": "Structured API failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "hint"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "hint": {
                "type": "string"
              }
            }
          }
        },
        "x-kiln-rust-type": "ApiError",
        "x-kiln-field-schema-status": "complete"
      },
...
      "TerminalStatusResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/TerminalStatusResponse",
        "x-kiln-rust-type": "TerminalStatusResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
      "WebSocketUpgrade": {
        "type": "string",
        "format": "binary",
        "description": "WebSocket upgrade transport body.",
        "x-kiln-rust-type": "WebSocketUpgrade",
        "x-kiln-field-schema-status": "complete"
      },
...
    }
  },
...
}